+++ CHANGELOG.rst
@@ -1,8 +1,16 @@
 
-0.1.1 (unreleased)
+0.2.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+**Features**
+
+- HashRedisMixin to store all objects on a single hash.
+- ``delete_all()`` and ``delete_filter()`` methods.
+
+**Bug Fixes**
+
+- Allow non-string id's to be used as primary keys.
+- Crash on ``all()`` with empty lists.
 
 
 0.1.0 (2017-09-12)